chore(cmd): :truck: Rename emoji to info emoji
